requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# JWT
python-jose[cryptography]==3.3.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1
# SQLite (로컬 테스트용)

# Password hashing
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Meilisearch
meilisearch==0.32.0

# HTTP Client
httpx==0.25.2

# AWS SDK
boto3==1.34.0

# Environment
python-dotenv==1.0.0
